refactor metabase-lib query utils
