+++ enterprise/frontend/src/embedding-sdk-package/hooks/public/use-metabase-query/metabase-lib-query-lib.ts
@@ -1,0 +1,14 @@
+// eslint-disable-next-line metabase/no-external-references-for-sdk-package-code
+import * as Lib from "metabase-lib";
+
+export { Lib };
+
+export type Aggregable = Lib.Aggregable;
+export type AggregationClause = Lib.AggregationClause;
+export type ColumnMetadata = Lib.ColumnMetadata;
+export type ExpressionClause = Lib.ExpressionClause;
+export type MeasureMetadata = Lib.MeasureMetadata;
+export type MetadataInput = Lib.Metadata;
+export type MetricMetadata = Lib.MetricMetadata;
+export type Query = Lib.Query;
+export type SegmentMetadata = Lib.SegmentMetadata;
