+++ enterprise/frontend/src/embedding-sdk-package/hooks/public/use-metabase-query/metabase-lib-query-normalization.ts
@@ -1,0 +1,69 @@
+import type { Aggregation, StructuredDatasetQuery } from "metabase-types/api";
+
+export function normalizeMetricAggregations(
+  datasetQuery: StructuredDatasetQuery,
+): StructuredDatasetQuery {
+  return {
+    ...datasetQuery,
+    query: {
+      ...datasetQuery.query,
+      aggregation: datasetQuery.query.aggregation?.map((aggregation) => {
+        if (
+          Array.isArray(aggregation) &&
+          aggregation[0] === "measure" &&
+          aggregation.length === 2
+        ) {
+          return ["measure", {}, aggregation[1]] as Aggregation;
+        }
+
+        return aggregation;
+      }),
+    },
+  };
+}
+
+export function normalizeDatasetQuery(
+  datasetQuery: StructuredDatasetQuery,
+): StructuredDatasetQuery {
+  return {
+    ...datasetQuery,
+    parameters: datasetQuery.parameters ?? [],
+    query: stripFieldRefBaseTypes(datasetQuery.query),
+  };
+}
+
+function stripFieldRefBaseTypes<TValue>(value: TValue): TValue {
+  if (Array.isArray(value)) {
+    if (
+      value[0] === "field" &&
+      typeof value[2] === "object" &&
+      value[2] != null &&
+      !Array.isArray(value[2])
+    ) {
+      const { "base-type": _baseType, ...options } = value[2] as Record<
+        string,
+        unknown
+      >;
+
+      return [
+        value[0],
+        value[1],
+        stripFieldRefBaseTypes(options),
+        ...value.slice(3).map(stripFieldRefBaseTypes),
+      ] as TValue;
+    }
+
+    return value.map(stripFieldRefBaseTypes) as TValue;
+  }
+
+  if (typeof value === "object" && value != null) {
+    return Object.fromEntries(
+      Object.entries(value).map(([key, childValue]) => [
+        key,
+        stripFieldRefBaseTypes(childValue),
+      ]),
+    ) as TValue;
+  }
+
+  return value;
+}
